refactor: move error classes from sdk to core to prevent circular dependencies
Key Changes:
- エラークラス（MetatellError基底クラスと派生クラス）をbot-coreパッケージに移動
- bot-sdkは後方互換性のためbot-coreからエラークラスを再エクスポート
- bot-realtimeの依存をbot-sdkからbot-coreに変更し、循環参照を解消

Benefits:
- SDKがファサードとしてrealtimeパッケージを呼び出す際の循環参照を防止
- パッケージ間の責任分離を明確化
- 既存のAPIとの後方互換性を維持
