+++ packages/core/package.json
@@ -23,8 +23,8 @@
     "@livekit/agents": "1.0.0",
     "@types/uuid": "10.0.0",
     "phoenix": "1.8.1",
-    "uuid": "12.0.0",
-    "ws": "8.18.3"
+    "uuid": "^12.0.0",
+    "ws": "^8.18.3"
   },
   "peerDependencies": {
     "@types/node": "*",
