+++ packages/realtime/package.json
@@ -28,7 +28,7 @@
   "license": "MIT",
   "dependencies": {
     "@livekit/rtc-node": "0.13.18",
-    "@metatell/bot-sdk": "workspace:*"
+    "@metatell/bot-core": "workspace:*"
   },
   "repository": {
     "type": "git",
